+++ Assets/Prefabs/Interactables/PingRing.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 3108945999165767123}
   - component: {fileID: 572422157530802717}
   - component: {fileID: 4773396187359916887}
+  - component: {fileID: 2702421996825939494}
   m_Layer: 0
   m_Name: PingRing
   m_TagString: Untagged
@@ -3008,3 +3009,15 @@ MeshCollider:
   m_Convex: 1
   m_CookingOptions: 30
   m_Mesh: {fileID: 0}
+--- !u!114 &2702421996825939494
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1645977202193189007}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bb6eda636b93f8f40b28d21e357bae21, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
