+++ Assets/Scripts/Interactables/NoisePinger.cs.meta
@@ -1,9 +1,11 @@
 fileFormatVersion: 2
-guid: bb6eda636b93f8f40b28d21e357bae21
+guid: 73547dfa966127f4096953b30953478b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
-  defaultReferences: []
+  defaultReferences:
+  - pingPrefab: {fileID: 1645977202193189007, guid: f1b3677185bfc3a489e4bd4feb07ace0,
+      type: 3}
   executionOrder: 0
   icon: {instanceID: 0}
   userData:
